Implement Schmitt's regularisation for TDVP (#1306)
This PR Implements the SNR regulariser by Markus Schmitt and Heyl
(https://journals.aps.org/prl/abstract/10.1103/PhysRevLett.125.100503).

The implementation works and has been tested in the last few weeks by
several persons, however I'm not happy with the interface of this
(namely, having a different driver).

Probably we would need to create some sort of 'preconditioner' interface
for the regularisation we want to use, but for now I think this is good
enough and would benefit from being shipped in NetKet